+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this "vscode-msr" extension will be documented in this file.
 
+## [2.1.26]
+
+- Improve class searching of C# code: filtering results by namespaces.
+- Added `msr.default.preferSearchingSpeedOverPrecision` to support precision for small projects.
+
 ## [2.1.25]
 
 - Robust escaping selected text when running search commands by menus.